Preserve dormant enemy previous-state words
RHArtificialMalignity leaves mPreviousState and mPreviousSubstate uninitialized in its constructor but serializes both complete four-byte enum storage words. Profile 011 therefore legitimately contains previous_state=72 while the pair is dormant.

Store the pair as raw i32 words, overwrite both at the Original Charly transition, and validate them only in the sole branch that semantically restores them. This keeps corrupt live state loud while allowing inert constructor bytes to round-trip exactly.

+++ crates/robin_engine/src/ai_enemy/battle.rs
@@ -1321,8 +1321,8 @@ impl EnemyAi {
                     }
 
                     // Only on first battle decision entry.
-                    if self.previous_substate == Substate::AttackingReactiontime
-                        || self.previous_substate == Substate::AttackingReactiontimeRunning
+                    if self.previous_substate == Substate::AttackingReactiontime as i32
+                        || self.previous_substate == Substate::AttackingReactiontimeRunning as i32
                     {
                         if self.is_vip {
                             self.base.say(Remark::VipProudDontFight);
